+++ trilogy-client/renderer/renderer.hpp
@@ -21,7 +21,7 @@ extern LRESULT ImGui_ImplWin32_WndProcHandler(HWND hWnd, UINT msg, WPARAM wParam
 class c_renderer : public c_singleton<c_renderer> {
 private:
 	using renderer_present_t = int32_t(__stdcall*)(IDXGISwapChain*, uint32_t, uint32_t);
-	using renderer_resize_t = int32_t(__stdcall*)(IDXGISwapChain*, uint32_t, uint32_t, uint32_t, DXGI_FORMAT, uint32_t);
+	using renderer_resize_t = int32_t(__stdcall*)(IDXGISwapChain*, uint32_t, FLOAT, FLOAT, DXGI_FORMAT, uint32_t);
 	using renderer_wndproc_t = int64_t(__stdcall*)(HWND, uint32_t, WPARAM, LPARAM);
 public:
 	renderer_present_t o_present;
